split BindingManager.ts
