+++ src/binding/BindingManager.ts
@@ -2,23 +2,22 @@ import { BaseComponent } from "components/base";
 import { AnchorGroupSchema, AnchorProxy, AnchorIdentifer } from "types/anchors";
 import { Field } from "vega-lite/build/src/channeldef";
 import { TopLevelSpec, UnitSpec } from "vega-lite/build/src/spec";
-import { compilationContext, deduplicateById, validateComponent, removeUndefinedInSpec ,logComponentInfo, detectAndMergeSuperNodes,  resolveAnchorValue} from "./binding";
-import {getProxyAnchor} from '../utils/anchorProxy';
+import { compilationContext, deduplicateById, validateComponent, removeUndefinedInSpec, logComponentInfo, detectAndMergeSuperNodes, resolveAnchorValue } from "./binding";
+import { getProxyAnchor } from '../utils/anchorProxy';
 import { VariableParameter } from "vega-lite/build/src/parameter";
-import {TopLevelSelectionParameter} from "vega-lite/build/src/selection"
+import { TopLevelSelectionParameter } from "vega-lite/build/src/selection"
 import { getChannelFromEncoding } from "../utils/anchorGeneration/rectAnchors";
-import {Edge} from "./binding"
+import { BindingEdge, GraphManager } from "./GraphManager";
+import {SpecCompiler} from "./SpecCompiler";
+
 type Parameter = VariableParameter | TopLevelSelectionParameter
 
 
-interface BindingNode {
-    id: string;
-    type: string;
-}
 
-export interface BindingEdge {
-    source: { nodeId: string; anchorId: string; };
-    target: { nodeId: string; anchorId: string; };
+export interface VirtualBindingEdge {
+    channel: string;
+    value: any;
+    source: 'context' | 'baseContext' | 'generated';
 }
 
 interface Binding {
@@ -28,29 +27,29 @@ interface Binding {
     targetAnchor: string;
 }
 
-interface BindingGraph {
-    nodes: Map<string, BindingNode>;
-    edges: BindingEdge[];
-}
-
-export interface VirtualBindingEdge {
-    channel: string;
-    value: any;
-    source: 'context' | 'baseContext' | 'generated';
-}
-
-
 export class BindingManager {
     private static instance: BindingManager;
-
     private graphManager: GraphManager;
     private specCompiler: SpecCompiler;
+
     private components: Map<string, BaseComponent> = new Map();
-    private virtualBindings: Map<string, VirtualBindingEdge> = new Map();
     private bindings: Binding[] = [];
+    private virtualBindings: Map<string, VirtualBindingEdge> = new Map();
 
+    private constructor() {
+        // Initialize dependencies lazily
+        this.graphManager = new GraphManager(() => this);
+        this.specCompiler = new SpecCompiler(this.graphManager, () => this);
+    }
 
+    public static getInstance(): BindingManager {
+        if (!BindingManager.instance) {
+            BindingManager.instance = new BindingManager();
+        }
+        return BindingManager.instance;
+    }
 
+    // Components
     public getComponents(): Map<string, BaseComponent> {
         return this.components;
     }
@@ -63,501 +62,55 @@ export class BindingManager {
         return component;
     }
 
-    public getGraphManager(): GraphManager {
-        return this.graphManager;
-    }
-
-    public compile(fromComponentId: string): TopLevelSpec {
-        return this.specCompiler.compile(fromComponentId);
-    }
-
-   
-    private constructor() {
-        // Initialize dependencies lazily
-        this.graphManager = new GraphManager( ()=>this);
-        this.specCompiler = new SpecCompiler(this.graphManager, ()=>this);
-        console.log('BindingManager initialized',this)
-    }
-
-    public static getInstance(): BindingManager {
-        if (!BindingManager.instance) {
-            BindingManager.instance = new BindingManager();
-        }
-        return BindingManager.instance;
-    }
-
     public addComponent(component: BaseComponent): void {
         this.components.set(component.id, component);
     }
 
-    public addVirtualBinding(channel: string, virtualBinding: VirtualBindingEdge): void {
-        this.virtualBindings.set(channel, virtualBinding);
-    }
 
-    public addBinding(sourceId: string, targetId: string, sourceAnchor: string, targetAnchor: string): void {
-        // Add the original binding
-        this.bindings.push({ sourceId, targetId, sourceAnchor, targetAnchor });
-                
-    };
-       
+    // Bindings 
     public getBindings(): Binding[] {
         return this.bindings;
     }
-    
 
-    public getBindingsForComponent(componentId: string): Binding[] {
-        return this.bindings.filter(binding =>
-            binding.sourceId === componentId || binding.targetId === componentId
-        );
+    public getBindingsForComponent(componentId: string, type?: 'source' | 'target' | 'both'): Binding[] {
+        return this.bindings.filter(binding => {
+            switch (type) {
+                case 'source':
+                    return binding.sourceId === componentId;
+                case 'target':
+                    return binding.targetId === componentId;
+                case 'both':
+                default:
+                    return binding.sourceId === componentId || binding.targetId === componentId;
+            }
+        });
     }
-    public getTargetBindingsForComponent(componentId: string): Binding[] {
-        return this.bindings.filter(binding =>
-            binding.targetId === componentId
-        );
-    }
-    public getSourceBindingsForComponent(componentId: string): Binding[] {
-        return this.bindings.filter(binding =>
-            binding.sourceId === componentId
-        );
+
+    public addBinding(sourceId: string, targetId: string, sourceAnchor: string, targetAnchor: string): void {
+        this.bindings.push({ sourceId, targetId, sourceAnchor, targetAnchor });
+    };
+
+
+    /* 
+    Virtual Bindings are bindings that come from configurations, they pass in data
+    but they do not have a source component. Often, they are used to set initial values
+    */
+    public addVirtualBinding(channel: string, virtualBinding: VirtualBindingEdge): void {
+        this.virtualBindings.set(channel, virtualBinding);
     }
 
     public getVirtualBindings(): Map<string, VirtualBindingEdge> {
         return this.virtualBindings;
     }
-}
 
-
-export class GraphManager {
-
-    constructor(
-        private getBindingManager: () => BindingManager // Getter for BindingManager
-    ) {}
-
-    public generateBindingGraph(startComponentId: string): BindingGraph {
-        const nodes = new Map<string, BindingNode>();
-        const edges: BindingEdge[] = [];
-        const visited = new Set<string>();
-        const bindingManager = this.getBindingManager(); // Access BindingManager when needed
-        
-
-        const addNode = (component: BaseComponent) => {
-            if (!nodes.has(component.id)) {
-                nodes.set(component.id, {
-                    id: component.id,
-                    type: component.constructor.name,
-                });
-            }
-        };
-
-        const traverse = (componentId: string) => {
-            if (visited.has(componentId)) return;
-            visited.add(componentId);
-
-            const component = bindingManager.getComponent(componentId);
-            if (!component) return;
-
-            addNode(component);
-
-            bindingManager.getSourceBindingsForComponent(componentId).forEach(binding => {
-                const { sourceId, targetId, sourceAnchor, targetAnchor } = binding;
-
-                [sourceId, targetId].forEach(id => {
-                    const comp = bindingManager.getComponent(id);
-                    if (comp) addNode(comp);
-                });
-
-                edges.push({
-                    source: { nodeId: sourceId, anchorId: sourceAnchor },
-                    target: { nodeId: targetId, anchorId: targetAnchor }
-                });
-
-                traverse(sourceId);
-                traverse(targetId);
-            });
-        };
-
-        traverse(startComponentId);
-        return { nodes, edges };
-    }
-
-    private superNodes: Map<string, string> = new Map();
-    public setSuperNodes(superNodeMap: Map<string, string>): void {
-        this.superNodes = superNodeMap;
-    }
-
-    public getSuperNodes(): Map<string, string> {
-        return this.superNodes;
-    }
-
-    public printGraph(startComponentId: string): void {
-        const { nodes, edges } = this.generateBindingGraph(startComponentId);
-
-        nodes.forEach(node =>
-            console.log(`  ${node.id} (${node.type})\n}`)
-        );
-
-        console.log('\nEdges:');
-        edges.forEach(edge =>
-            console.log(`  ${edge.source.nodeId}.${edge.source.anchorId} -> ${edge.target.nodeId}.${edge.target.anchorId}`)
-        );
-    }
-}
-
-interface AnchorEdge {
-    originalEdge: BindingEdge;
-    anchorProxy: AnchorProxy;
-}
-
-export type Edge = AnchorEdge | VirtualBindingEdge;
-
-
-export function expandGroupAnchors(
-    edge: AnchorEdge,
-    component: BaseComponent
-): AnchorEdge[] {
-    const { originalEdge, anchorProxy } = edge;
-    const schema = anchorProxy.anchorSchema;
-    
-    if (schema.type === 'group') {
-        console.log('expanding group anchors',schema.children)
-
-        return schema.children.map(childId => ({
-            originalEdge,
-            anchorProxy: component.getAnchor(childId)
-        }));
-    }
-    return [edge];
-}
-export class SpecCompiler {
-    constructor(
-        private graphManager: GraphManager,
-        private getBindingManager: () => BindingManager // Getter for BindingManager
-    ) {}
-
-    private hierarchyEdges: Map<string, BindingEdge[]> = new Map();
-
-
-    // note: component Id will always be called from the root component
     public compile(fromComponentId: string): TopLevelSpec {
-
-        const rootComponent = this.getBindingManager().getComponent(fromComponentId);
-        if (!rootComponent) {
-            throw new Error(`Component "${fromComponentId}" not found.`);
-        }
-
-        // specific binding graph for this tree
-        let bindingGraph = this.graphManager.generateBindingGraph(rootComponent.id);
-       
-        // Compile the updated graph
-        const compiledSpecs = this.compileBindingGraph(bindingGraph);
-
-        //const compiledSpecs = this.compileBindingGraph(bindingGraph);
-        const mergedSpec = mergeSpecs(compiledSpecs,  rootComponent.id);
-
-        return removeUndefinedInSpec(mergedSpec);
+        return this.specCompiler.compile(fromComponentId);
     }
-    
-    
-
-    private compileBindingGraph(bindingGraph: BindingGraph): Partial<UnitSpec<Field>>[] {
-        const { nodes, edges } = bindingGraph;
-        const compiledComponents: Partial<UnitSpec<Field>>[] = [];
-
-        const superNodeMap : Map<string, string>= detectAndMergeSuperNodes(edges);
-        this.graphManager.setSuperNodes(superNodeMap);
-
-
-
-
-        for (const node of nodes.values()) {
-            const compiledNode = this.compileNode(node, edges);
-            compiledComponents.push(compiledNode);
-        }
-
-        return compiledComponents;
-    }
-
-    
-    private compileNode(node: BindingNode, graphEdges: BindingEdge[]): Partial<UnitSpec<Field>> {
-
-        const filteredEdges = graphEdges.filter(edge => edge.target.nodeId === node.id )
-
-        // maybe instead of turn edges into anchors, we should just keep edges, and add anchor property
-        const incomingAnchors : AnchorEdge[] = this.prepareEdges(filteredEdges)
-
-        // Okay, at this point you kept the edge with the anchor itself. Now you need to pipe those changes through. 
-        // this will be nice because you can use the edge targetId to group now.
-        // then within that, you just pipe the 
-
-
-        //keeping a VSCode History Log- all of the comments you make to yourself, get logged. and the the compiler thinks of if 
-        // you're talking to it, and then describes all fo th changes in a really nice git history. 
-
-        // helpful for understanding 
-        
-
-
-
-
-        const anchorProxies = incomingAnchors.map(edge => edge.anchorProxy);
-        //then, as we compile the node, we pass in the edges and it groups on edge.target.anchorId (then compiles)
-
-
-        const superNodeMap : Map<string, string>= this.graphManager.getSuperNodes();
-
-        let compilationContext = {nodeId: superNodeMap.get(node.id) || node.id};
-
-        let component = this.getBindingManager().getComponent(node.id);
-
-        
-
-        compilationContext = this.buildPersonalizedCompilationContext(component, incomingAnchors, compilationContext);
-
-        compilationContext = this.scalePropagation(node.id,compilationContext);
-
-        console.log('compilationContext',compilationContext)
-
-
-        // at this point compilationContext will have a grouped and ordered list of corresponding values. 
-        
-
-        const compiledSpec = this.compileComponentWithContext(node.id, compilationContext);
-        return compiledSpec;
-    }
-
-    private findRootNode(nodeId: string): string {
-        const bindingManager = this.getBindingManager();
-        // search up the binding graph until we find the root node
-        let currentId = nodeId;
-        // while (bindingManager.getSourceBindingsForComponent(currentId).length > 0) {
-        //     console.log('currentId',currentId)
-        //     currentId = bindingManager.getTargetBindingsForComponent(currentId)[0].sourceId;
-        // }
-        return currentId;
-    }
-
-    private scalePropagation(nodeId: string, compilationContext: compilationContext): compilationContext {
-        // Find root node by traversing up the binding graph
-        const rootNodeId = 'node_3' //this.findRootNode(nodeId);
-        if (!rootNodeId) return compilationContext;
-
-        const rootComponent = this.getBindingManager().getComponent(rootNodeId);
-        if (!rootComponent) return compilationContext;
-
-        console.log('rootComponent',rootComponent)
-        // For each key in compilation context
-        for (const [key, value] of Object.entries(compilationContext)) {
-            if (key === 'nodeId') continue;
-            // Get all anchors from root component
-            const rootAnchors = rootComponent.getAnchors();
-            
-            // Try to find matching anchor by id or channel name
-            const rootAnchor = rootAnchors.find(anchor => {
-                const anchorId = anchor.id.anchorId;
-                const channelName = getChannelFromEncoding(key);
-
-                return anchorId === key || channelName === anchorId;
-            });
-            if (!rootAnchor) continue;
-
-            // Compile the root anchor's value
-            const compiledValue = rootAnchor.compile?.(rootNodeId);
-            if (!compiledValue) continue;
-            // Add scale and scaleType from root to current context if they exist
-            if (compiledValue.value?.scale) {
-                compilationContext[key].scale = compiledValue.value.scale;
-            }
-            if (compiledValue.value?.scaleType) {
-                compilationContext[key].scaleType = compiledValue.value.scaleType;
-            }
-        }
-        return compilationContext;
-    }
-
-    private prepareEdges(graphEdges: BindingEdge[]): AnchorEdge[] {
-
-        function isCompatible(sourceAnchorId:string,targetAnchor:string){
-            return getChannelFromEncoding(sourceAnchorId) == getChannelFromEncoding(targetAnchor)
-        }
-        function expandAllAnchors(edge: BindingEdge, source: BaseComponent, target: BaseComponent): BindingEdge[] {
-            const getAnchors = (component: BaseComponent, anchorId: string) => 
-                anchorId === '_all' 
-                    ? [...component.getAnchors().values()].map(a => a.id.anchorId)
-                    : [anchorId];
-        
-            const sourceAnchors = getAnchors(source, edge.source.anchorId);
-            const targetAnchors = getAnchors(target, edge.target.anchorId);
-        
-            return sourceAnchors.flatMap(sourceAnchor => 
-                targetAnchors
-                    .filter(targetAnchor => {
-                        const sourceType = source.getAnchor(sourceAnchor)?.anchorSchema.type;
-                        const targetType = target.getAnchor(targetAnchor)?.anchorSchema.type;
-                        return sourceType === targetType && isCompatible(sourceAnchor,targetAnchor);
-                    })
-                    .map(targetAnchor => ({
-                        source: { nodeId: edge.source.nodeId, anchorId: sourceAnchor },
-                        target: { nodeId: edge.target.nodeId, anchorId: targetAnchor }
-                    }))
-            );
-        }
-        // new issue: we are now dropping baseContext data rather than just retrning null (if nothing is provided.)
-
-        return graphEdges.flatMap(edge => {
-            const sourceComponent = BindingManager.getInstance().getComponent(edge.source.nodeId);
-            if (!sourceComponent) {
-                throw new Error(`Source component ${edge.source.nodeId} not found`);
-            }
-
-            // Handle '_all' anchor expansion
-            if (edge.source.anchorId === '_all' || edge.target.anchorId === '_all') {
-                const targetComponent = BindingManager.getInstance().getComponent(edge.target.nodeId);
-                if (!targetComponent) {
-                    throw new Error(`Target component ${edge.target.nodeId} not found`);
-                }
-                
-                return expandAllAnchors(edge, sourceComponent, targetComponent)
-                    .flatMap(newEdge => {
-                        const anchorProxy = sourceComponent.getAnchor(newEdge.source.anchorId);
-                        if (!anchorProxy) return [];
-                        return expandGroupAnchors({
-                            originalEdge: newEdge,
-                            anchorProxy
-                        }, sourceComponent);
-                    });
-            }
-
-            // Handle regular edges
-            const anchorProxy = sourceComponent.getAnchor(edge.source.anchorId);
-            if (!anchorProxy) return [];
-            return expandGroupAnchors({
-                originalEdge: edge,
-                anchorProxy
-            }, sourceComponent);
-        });
-    }
-   
-    // private prepareEdges(graphEdges: BindingEdge[]): AnchorProxy[] {
-    //     const anchorProxies = graphEdges
-    //         .map(edge => getProxyAnchor(edge, this.getBindingManager().getComponent(edge.source.nodeId)));
-
-    //     const expandedEdges = anchorProxies.flatMap(anchorProxy =>
-    //         expandGroupAnchors(anchorProxy, anchorProxy.component)
-    //     );
-
-    //     return deduplicateById(expandedEdges);
-    // }
-
-
-    private buildPersonalizedCompilationContext(component: BaseComponent, edges: AnchorEdge[], compilationContext: compilationContext): compilationContext {
-        // for each of component's anchors, we need to produce a group of the edges that are associated with that anchor
-        
-        const anchors = component.getAnchors();
-       
-
-        // group based on targetAnchorId
-        
-        const groupedEdges = edges.reduce((acc, edge) => {
-            const targetAnchorId = edge.originalEdge.target.anchorId;
-            if (!acc.has(targetAnchorId)) {
-                acc.set(targetAnchorId, []);
-            }
-            acc.get(targetAnchorId)?.push(edge);
-            return acc;
-        }, new Map<string, Edge[]>());
-
-
-        // filter out incompatible edges
-        // for each property make sure that the 
-
-        
-
-
-        // this.getBindingManager().getVirtualBindings().forEach((virtualBinding, channel) => {
-        //     if (!groupedEdges.has(channel)) {
-        //         groupedEdges.set(channel, []);
-        //     }
-        //     groupedEdges.get(channel)?.push(virtualBinding);
-        // });
-        
-
-
-
-
-        // now for each of the anchorMatchedEdges, we need to resolve the value of the edges
-        for (const [anchorId, edges] of groupedEdges.entries()) {
-            const superNodeMap : Map<string, string>= this.graphManager.getSuperNodes();
-            const resolvedValue = resolveAnchorValue(edges, superNodeMap);
-            compilationContext[anchorId] = resolvedValue;
-        }
-
-
-            // Add scale propagation from parent components
-        const hierarchyAncestors = this.getHierarchyAncestors(compilationContext.nodeId);
-
-        hierarchyAncestors.forEach(ancestorId => {
-            const ancestorSpec = this.compileComponentWithContext(ancestorId, compilationContext);
-            Object.entries(ancestorSpec.encoding || {}).forEach(([channel, enc]) => {
-                if (!compilationContext[channel]?.scale ) {
-                    compilationContext[channel] = {
-                        ...compilationContext[channel],
-                        scale: enc?.scale,
-                        scaleType: enc?.scale?.type
-                    };
-                }
-            });
-        });
-
-        
-        return compilationContext
-
-        // TODO, channel edges have second priorty. 
-        // for example, if a brush.left : drag_line, then we need to resolve brush.x1 (left), to drag_line.x, 
-        // so if none of the edges have a direct map, then we use channel edges to resolve the value
-        
-
-
-
-        //console.log('personalized anchorMatchedEdges',anchorMatchedEdges)
-        // console.log('personalized filteredChannelEdges',filteredChannelEdges)
-
-
-        // we need to then resolve the value of each of the edges in the group
-
-        // we need to then add the resolved value to the compilationContext
-    }
-
-
-    private getHierarchyAncestors(nodeId: string): string[] {
-        const ancestors: string[] = [];
-        let currentId = nodeId;
-        
-        while (this.hierarchyEdges.has(currentId)) {
-            const edges = this.hierarchyEdges.get(currentId)!;
-            edges.forEach(edge => {
-                ancestors.push(edge.source.nodeId);
-                currentId = edge.source.nodeId;
-            });
-        }
-        
-        return ancestors;
-    }
-
-
-    private compileComponentWithContext(nodeId: string, context: compilationContext): Partial<UnitSpec<Field>> {
-        const component = this.getBindingManager().getComponent(nodeId);
-        if (!component) {
-            throw new Error(`Component "${nodeId}" not found.`);
-        }
-        validateComponent(component, nodeId);
-        
-        return component.compileComponent(context);
-    }
-
-    
 }
 
-function  mergeSpecs(specs: Partial<UnitSpec<Field>>[], rootComponentId: string): TopLevelSpec {
+
+
+function mergeSpecs(specs: Partial<UnitSpec<Field>>[], rootComponentId: string): TopLevelSpec {
     // Helper to check if spec has layer/mark
     const hasLayerOrMark = (spec: any) => {
         return spec.layer || spec.mark;
@@ -595,21 +148,21 @@ function  mergeSpecs(specs: Partial<UnitSpec<Field>>[], rootComponentId: string)
     // Function to find base chart layer by rootComponentId
     const findBaseChartLayer = (obj: any): any | null => {
         if (!obj || typeof obj !== 'object') return null;
-        
+
         if (obj.name === rootComponentId) return obj;
-        
+
         for (const value of Object.values(obj)) {
             const result = findBaseChartLayer(value);
             if (result) return result;
         }
-        
+
         return null;
     };
 
     // Function to move params to top level
     const moveParamsToTop = (obj: any) => {
         if (!obj || typeof obj !== 'object') return;
-        
+
         if (obj.params && Array.isArray(obj.params)) {
             // Split params into select and non-select
             const [selectParamsArr, nonSelectParamsArr] = obj.params.reduce(
@@ -623,7 +176,7 @@ function  mergeSpecs(specs: Partial<UnitSpec<Field>>[], rootComponentId: string)
                 },
                 [[], []]
             );
-            
+
             // Add to respective arrays
             if (nonSelectParamsArr.length > 0) {
                 params.push(...nonSelectParamsArr);
@@ -631,11 +184,11 @@ function  mergeSpecs(specs: Partial<UnitSpec<Field>>[], rootComponentId: string)
             if (selectParamsArr.length > 0) {
                 selectParams.push(...selectParamsArr);
             }
-            
+
             // Remove params from original location
             delete obj.params;
         }
-        
+
         Object.values(obj).forEach(value => {
             moveParamsToTop(value);
         });
@@ -659,11 +212,11 @@ function  mergeSpecs(specs: Partial<UnitSpec<Field>>[], rootComponentId: string)
 
 function mergeParams(params: any[]): any[] {
     const paramsByName = new Map<string, any>();
-    
+
     // Group params by name
     params.forEach(param => {
         if (!param.name) return;
-        
+
         if (paramsByName.has(param.name)) {
             // Merge with existing param
             const existing = paramsByName.get(param.name);
